refactoring: new package and structure for CF/DF related validations
